Orchestration/Digicert AP Redirector/requirements.txt
cloudshell-orch-core>=1.7.0.0,<1.8.0.0
requests==2.26.0